add ctrlWrapper decorator
