+++ decorators/ctrlWrapper.js
@@ -1,0 +1,14 @@
+const ctrlWrapper = ctrl => {
+    const func = async (req, res, next) => {
+        try {
+            await ctrl(req, res, next);
+        }
+        catch (error) {
+            next(error);
+        }
+    }
+
+    return func;
+}
+
+module.exports = ctrlWrapper;
