+++ decorators/index.js
@@ -1,0 +1,6 @@
+const ctrlWrapper = require('./ctrlWrapper');
+
+
+module.exports = {
+    ctrlWrapper,
+}
