+++ routes/api/contacts.js
@@ -1,5 +1,5 @@
 const express = require('express')
-const ctrl = require('../../controllers/contacts-controller')
+const ctrl = require('../../controllers/contactsController')
 
 const router = express.Router()
